recruiter-agent/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiomysql==0.2.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asn1crypto==1.5.1
async-timeout==5.0.1
asyncpg==0.29.0
attrs==25.1.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
boto3==1.36.26
botocore==1.36.26
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
cryptography==42.0.8
distro==1.9.0
dotenv==0.9.9
duckdb==1.2.0
fastapi==0.115.8
fastapi-inertia==1.0.3
filelock==3.17.0
frozenlist==1.5.0
google==3.0.0
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
gspread==6.1.4
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-core==0.3.37
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langsmith==0.3.10
MarkupSafe==3.0.2
morph-data==0.2.0
morph-lib==0.0.1
multidict==6.1.0
narwhals==1.29.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.64.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
platformdirs==4.3.6
plotly==6.0.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyMySQL==1.1.1
PyNaCl==1.5.0
pyOpenSSL==24.3.0
pyparsing==3.2.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
python-tds==1.16.0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.13.2
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==1.4.52
sqlalchemy-redshift==0.8.14
sqlalchemy_pytds==1.0.2
sqlglot==25.34.1
sshtunnel==0.4.0
starlette==0.41.3
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.9.0
toml==0.10.2
tomlkit==0.13.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.32.1
yarl==1.18.3
zstandard==0.23.0
